requirements.txt
pytest==6.2.4
pytest-mock==3.6.1
flake8==3.9.2
mypy==0.910
PyJWT==2.1.0
Flask==2.0.1